api/records: Use Redis instead of DB
