+++ src/api/mod.rs
@@ -2,5 +2,4 @@ pub mod cors_proxy;
 pub mod ddb;
 pub mod live;
 pub mod records;
-pub mod redis_records;
 pub mod status;
